+++ CHANGELOG.md
@@ -18,6 +18,7 @@ contents to a new release section (e.g. _[0.x]_).
 - Solve elementary logarithmic inequalities. (PLUT-1135)
 - Solve logarithmic inequalities of form log(f) < log(g). (PLUT-1136)
 - Solve nested logarithmic equaqions. (PLUT-1140)
+- Solve nested logarithmic inequalities. (PLUT-1142)
 - SDK: Add support for exponent on log function. (PLUT-1127)
 
 ### Fixed